src/requirements.txt
# tensorflow == 1.12.0   # CPU Version of TensorFlow.
tensorflow-gpu  == 1.12.0  # GPU version of TensorFlow.